+++ database/migrations/2015_05_26_124719_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration {
 		{
             $table->increments('id');
             $table->integer('user_id')->unsigned();
-            $table->decimal('price', 8, 2);
+            $table->double('price', 8, 2);
             $table->string('currency');
             $table->integer('type_id')->unsigned();
             $table->string('comment');
