+++ database/migrations/2015_08_06_124230_create_dish_ratings_table.php
@@ -18,7 +18,7 @@ class CreateDishRatingsTable extends Migration {
 			$table->integer('user_id')->unsigned();
 			$table->integer('dish_id')->unsigned();
 			$table->integer('review_id')->unsigned();
-			$table->decimal('rating', 3, 1);
+			$table->double('rating', 2, 1);
 			$table->timestamps();
 		});
 	}
